+++ core/tracker.lua
@@ -65,6 +65,7 @@ M.reset_run = function ()
     M.external_caller   = nil
     M.external_callback = nil
     M.teleport_required = false
+    M.last_pick_entry   = nil
 end
 
 return M